Review_Site/requirements.txt
Django
djangorestframework
drf-spectacular
django-extensions
Pillow
django-cors-headers